+++ NEWS.md
@@ -1,6 +1,7 @@
 # ggpedigree (development version)
 # ==============================
 * Improved documentation of datasets included in the package
+* Added smarter warning message for tryCatch
 
 # ggpedigree 1.1.0.0
 * Saved raw excel data as csv to have a non-proprietary format for data storage